Use non-root user for containers

+++ images/jobs/keeper-upgrade.sh
@@ -2,7 +2,7 @@
 # -*- mode: sh; -*-
 
 # File: entrypoint.sh
-# Time-stamp: <2018-05-22 17:05:45>
+# Time-stamp: <2018-10-30 13:23:15>
 # Copyright (C) 2018 Gravitational Inc
 # Description:
 
@@ -19,22 +19,23 @@ do
     fi
 done
 
-# fix permissions for ssl
-cp -R /etc/ssl/cluster-default /etc/ssl/cluster-default-postgres
-chown -R stolon /etc/ssl/cluster-default-postgres
-chmod 0600 /etc/ssl/cluster-default-postgres/default-server-key.pem
-
-gosu stolon mkdir -p /stolon-data/postgres-new /stolon-data/upgrade-state
-cd /stolon-data/upgrade-state
-if [ ! -f /stolon-data/postgres-new/postgresql.conf ]; then
-    PGDATA=/stolon-data/postgres-new gosu stolon ${PGBINNEW}/initdb
-fi
-if [ ! -f /stolon-data/upgrade-state/delete_old_cluster.sh ]; then
-    gosu stolon ${PGBINNEW}/pg_upgrade -d /stolon-data/postgres -D /stolon-data/postgres-new
-    gosu stolon cp /stolon-data/postgres/pg_hba.conf /stolon-data/postgres-new/pg_hba.conf
-    gosu stolon rsync -av /stolon-data/postgres/conf.d /stolon-data/postgres-new/
-    gosu stolon rsync -av /stolon-data/postgres/postgresql-base.conf /stolon-data/postgres-new/
-    if [ -d /stolon-data/postgres-old-$PG_VERSION_OLD ]; then rm -rf /stolon-data/postgres-old-$PG_VERSION_OLD; fi
-    gosu stolon mv /stolon-data/postgres /stolon-data/postgres-old-$PG_VERSION_OLD
-    gosu stolon mv /stolon-data/postgres-new /stolon-data/postgres
+sed -i -e "s#\(ssl_cert_file = '\).*#\1/home/stolon/secrets/cluster-default/default-server.pem'#" /stolon-data/postgres/postgresql.conf
+sed -i -e "s#\(ssl_key_file = '\).*#\1/home/stolon/secrets/cluster-default/default-server-key.pem'#" /stolon-data/postgres/postgresql.conf
+sed -i -e "s#\(ssl_ca_file = '\).*#\1/home/stolon/secrets/cluster-ca/ca.pem'#" /stolon-data/postgres/postgresql.conf
+if [ $(cat /stolon-data/postgres/PG_VERSION) != $PG_VERSION_NEW ]
+then
+    mkdir -p /stolon-data/postgres-new /stolon-data/upgrade-state
+    cd /stolon-data/upgrade-state
+    if [ ! -f /stolon-data/postgres-new/postgresql.conf ]; then
+        PGDATA=/stolon-data/postgres-new ${PGBINNEW}/initdb
+    fi
+    if [ ! -f /stolon-data/upgrade-state/delete_old_cluster.sh ]; then
+        ${PGBINNEW}/pg_upgrade -d /stolon-data/postgres -D /stolon-data/postgres-new
+        cp /stolon-data/postgres/pg_hba.conf /stolon-data/postgres-new/pg_hba.conf
+        rsync -av /stolon-data/postgres/conf.d /stolon-data/postgres-new/
+        rsync -av /stolon-data/postgres/postgresql-base.conf /stolon-data/postgres-new/
+        if [ -d /stolon-data/postgres-old-$PG_VERSION_OLD ]; then rm -rf /stolon-data/postgres-old-$PG_VERSION_OLD; fi
+        mv /stolon-data/postgres /stolon-data/postgres-old-$PG_VERSION_OLD
+        mv /stolon-data/postgres-new /stolon-data/postgres
+    fi
 fi
